Moved import of thrift dependencies inside the conditional that needs them
Importing the thrift libraries is causing an exception to be raised when starting the reddit service. Since we don't care about using their activity tracking service (count of number of visitors) then we don't need to be able to load those libraries. This also encourages lazy loading to decrease overall memory requirement.